+++ phidget-sys/build.rs
@@ -27,6 +27,11 @@ fn main() {
     //      bindings file for the specific target.
     let tgt = env::var("TARGET").unwrap();
     println!("debug: Building for target: '{}'", tgt);
+    
+    // PHIDGET_ROOT should be set to point to the installation directory of phidgets (e.g. C:\Program Files\Phidgets\Phidget22)
+    if let Ok(phidget_libs) = env::var("PHIDGET_ROOT"){
+        println!("cargo:rustc-link-search={}", phidget_libs);
+    }
 
     #[cfg(target_os = "macos")]
     config_macos();