+++ src/bin/phidget.rs
@@ -1,4 +1,5 @@
-use phidget::Phidget;
+use phidget::{devices::{HumiditySensor, TemperatureSensor}, Phidget};
+
 use std::{thread, time::Duration};
 
 const TIMEOUT: Duration = Duration::from_millis(5000);
@@ -7,7 +8,7 @@ fn main() -> anyhow::Result<()> {
     println!("{}", phidget::library_version()?);
     println!("{}", phidget::library_version_number()?);
 
-    let mut hum_sensor = phidget::HumiditySensor::new();
+    let mut hum_sensor = HumiditySensor::new();
     phidget::phidget::set_on_attach_handler(&mut hum_sensor, |_| {
         println!("Humidity sensor attached");
     })?;
@@ -30,7 +31,7 @@ fn main() -> anyhow::Result<()> {
     let humidity = hum_sensor.humidity()?;
     println!("Humidity: {}", humidity);
 
-    let mut temp_sensor = phidget::TemperatureSensor::new();
+    let mut temp_sensor = TemperatureSensor::new();
     phidget::phidget::set_on_attach_handler(&mut temp_sensor, |_| {
         println!("Temperature sensor attached");
     })?;
@@ -53,12 +54,12 @@ fn main() -> anyhow::Result<()> {
     let temperature = temp_sensor.temperature()?;
     println!("Temperature: {}\n", temperature);
 
-    hum_sensor.set_on_humidity_change_handler(|_s: &phidget::HumiditySensor, humidity: f64| {
+    hum_sensor.set_on_humidity_change_handler(|_s: &HumiditySensor, humidity: f64| {
         println!("Humidity: {}", humidity);
     })?;
 
     temp_sensor.set_on_temperature_change_handler(
-        |_s: &phidget::TemperatureSensor, temperature: f64| {
+        |_s: &TemperatureSensor, temperature: f64| {
             println!("Temerature: {}", temperature);
         },
     )?;